Refactor + extension function(s) addition
